+++ Realm/RLMObjectBase.mm
@@ -93,7 +93,7 @@ static void maybeInitObjectSchemaForUnmanaged(RLMObjectBase *obj) {
 }
 
 static id coerceToObjectType(id obj, Class cls, RLMSchema *schema) {
-    if ([obj isKindOfClass:cls]) {
+    if ([obj isKindOfClass:cls] && (![(id)cls isEmbedded] || ![obj realm])) {
         return obj;
     }
     obj = RLMBridgeSwiftValue(obj) ?: obj;
@@ -537,6 +537,9 @@ id RLMObjectThaw(RLMObjectBase *obj) {
 }
 
 id RLMValidatedValueForProperty(id object, NSString *key, NSString *className) {
+    if ([object isKindOfClass:[RLMObjectBase class]] && ![[[object objectSchema] className] isEqualToString:className]) {
+        @throw RLMException(@"Invalid value: cannot initialize '%@' with value '%@'", className, object);
+    }
     @try {
         return [object valueForKey:key];
     }